Work around XCB error in Xlib cleanup
On modern versions of Xlib that are built on top of XCB, a failure
from glXCreateContext seems to leave the message queue in a broken
state that triggers an abort during global destruction.  The test
gets marked as a failure instead of 'skip'.  Work around this
problem by using POSIX::_exit(0) which skips perl's cleanup.